(cho) move type definations from config to types/index.ts file
